requirements.txt
streamlit
pymongo
google-generativeai
langchain
pandas
openai
python-dotenv
requests 
beautifulsoup4 
sentence-transformers 
faiss-cpu
langchain-community
PyPDF2
docx
mistralai